Improve file arch
